requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
pillow
replicate
openai 